refactor: separate the components of the chart
