+++ src/components/chart/styled.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const ContainerChart = styled.div`
+  border-bottom: 1px solid #ccc;
+  padding-top: 5%;
+  padding-bottom: 7%;
+`
